Rename routes dir to pages (#49)
